refactor: extract data collection logic
